Fix log processing scripts and increase client ops frequency to once every 50ms

+++ scripts/start-clients.sh
@@ -1,25 +1,33 @@
 #!/bin/bash
 
+if (($# != 6)); then
+	echo "Arguments: path_to_client, logs_dest, num_clients, num_servers, run_time, msg_length"
+	exit 1	
+fi
+
 trap 'kill $(jobs -p)' EXIT
 
 client_path=$1
 logs_dest=$2
 n=$((10#$3))
-t=$((10#$4))
-i=0
+ns=$((10#$4))
+t=$((10#$5))
+msg_len=$((10#$6))
+
+read servers_list
+read ports_list
+
 rm -rf $logs_dest && mkdir -p $logs_dest
+cd $logs_dest
+
+i=0
 while (($i < $n)); do
-	if (($i % 2 == 0)); then
-		cd $logs_dest
-		java -jar $client_path dryad01.ethz.ch 10002 $i $n > /dev/null &
-	else
-		cd $logs_dest
-		java -jar $client_path dryad02.ethz.ch 10002 $i $n > /dev/null &
-	fi
+	java -jar $client_path ${servers_list[$(($i % $ns))]} ${ports_list[$(($i % $ns))]} $i $n $msg_len > /dev/null &
 	i=$(($i + 1))
 done
 
-echo "Running a total of $n clients for a total of $t seconds."
+echo "Running $n clients for a total of $t seconds."
+echo "Message length set to $msg_len."
 echo "Log data will be available at $logs_dest"
 i=0
 while (($i < $t)); do
@@ -28,4 +36,4 @@ while (($i < $t)); do
 	sleep 1s
 done
 
-echo "Done.   
+echo "Done.                         "
